rejigger the filtering so we can run it in the tests
