+++ src/pudl/transform/__init__.py
@@ -71,6 +71,7 @@ from . import (
     eia_bulk_elec,
     eiaaeo,
     epacems,
+    ferc,
     ferc1,
     ferc714,
     gridpathratoolkit,
